Move controller advice to dedicated package and improve error messages
- Move NoCacheAdvice and GlobalExceptionHandler from controllers.rest to controllers.advice package
- Add descriptive messages to orElseThrow() calls in ResultService so failures are visible in logs
